[SEMVER-MAJOR] Reword log messages for clarity
Before this commit, we logged messages saying "unhandled error". This
was often creating confusion for our users, it seems like the error
was not handled at all, when in fact it was caught by the configured
error handler and converted into a proper HTTP error response.

In this commit, I am changing the messaging to use "request failed"
instead.

While this change is not affecting application developers, it may
break tooling processing logs if the tool relies on a specific format
of error messages. To be safe, I am labelling this change as breaking.

Signed-off-by: Miroslav Bajtoš <mbajtoss@gmail.com>